daemon/libnetwork: factor out iptables service
Factor the iptables-specific service mesh configuration out of the
common Linux service mesh implementation in preparation for adding an
nftables implementation.
